moved presence.py from base into binary_sensor package
