+++ custom_components/magic_areas/binary_sensor/presence.py
